Adsk Contrib - Improve RangeTransform
