+++ src/OpenColorIO/ops/Range/RangeOpCPU.h
@@ -2,8 +2,8 @@
 // Copyright Contributors to the OpenColorIO Project.
 
 
-#ifndef INCLUDED_OCIO_RANGEOP_CPU
-#define INCLUDED_OCIO_RANGEOP_CPU
+#ifndef INCLUDED_OCIO_RANGEOP_CPU_H
+#define INCLUDED_OCIO_RANGEOP_CPU_H
 
 
 #include <OpenColorIO/OpenColorIO.h>
