+++ src/OpenColorIO/ops/Range/RangeOpGPU.h
@@ -2,8 +2,8 @@
 // Copyright Contributors to the OpenColorIO Project.
 
 
-#ifndef INCLUDED_OCIO_RANGE_GPU
-#define INCLUDED_OCIO_RANGE_GPU
+#ifndef INCLUDED_OCIO_RANGEOP_GPU_H
+#define INCLUDED_OCIO_RANGEOP_GPU_H
 
 
 #include <OpenColorIO/OpenColorIO.h>
