+++ src/OpenColorIO/transforms/RangeTransform.cpp
@@ -7,79 +7,11 @@
 #include <OpenColorIO/OpenColorIO.h>
 
 #include "MathUtils.h"
-#include "ops/Range/RangeOpData.h"
+#include "RangeTransform.h"
 
 
 OCIO_NAMESPACE_ENTER
 {
-
-class RangeTransformImpl : public RangeTransform
-{
-public:
-    RangeTransformImpl() = default;
-    RangeTransformImpl(const RangeTransformImpl &) = delete;
-    RangeTransformImpl& operator=(const RangeTransformImpl &) = delete;
-    virtual ~RangeTransformImpl() = default;
-
-    TransformRcPtr createEditableCopy() const override;
-
-    TransformDirection getDirection() const override;
-    void setDirection(TransformDirection dir) override;
-
-    FormatMetadata & getFormatMetadata() override;
-    const FormatMetadata & getFormatMetadata() const override;
-
-    BitDepth getFileInputBitDepth() const override;
-    BitDepth getFileOutputBitDepth() const override;
-    void setFileInputBitDepth(BitDepth bitDepth) override;
-    void setFileOutputBitDepth(BitDepth bitDepth) override;
-
-    RangeStyle getStyle() const override;
-    void setStyle(RangeStyle style) override;
-
-    void validate() const override;
-
-    bool equals(const RangeTransform & other) const override;
-
-    void setMinInValue(double val) override;
-    double getMinInValue() const override;
-    bool hasMinInValue() const override;
-    void unsetMinInValue() override;
-
-    void setMaxInValue(double val) override;
-    double getMaxInValue() const override;
-    bool hasMaxInValue() const override;
-    void unsetMaxInValue() override;
-
-    void setMinOutValue(double val) override;
-    double getMinOutValue() const override;
-    bool hasMinOutValue() const override;
-    void unsetMinOutValue() override;
-
-    void setMaxOutValue(double val) override;
-    double getMaxOutValue() const override;
-    bool hasMaxOutValue() const override;
-    void unsetMaxOutValue() override;
-
-    RangeOpData & data() { return m_data; }
-    const RangeOpData & data() const { return m_data; }
-
-    static void deleter(RangeTransformImpl * t)
-    {
-        delete t;
-    }
-
-private:
-    TransformDirection m_direction = TRANSFORM_DIR_FORWARD;
-    RangeStyle  m_style = RANGE_CLAMP;
-    RangeOpData m_data;
-};
-
-
-
-///////////////////////////////////////////////////////////////////////////
-
-
 
 RangeTransformRcPtr RangeTransform::Create()
 {
@@ -267,107 +199,3 @@ std::ostream& operator<< (std::ostream & os, const RangeTransform & t)
 
 }
 OCIO_NAMESPACE_EXIT
-
-
-////////////////////////////////////////////////////////////////////////////////
-
-
-#ifdef OCIO_UNIT_TEST
-
-#include "UnitTest.h"
-
-namespace OCIO = OCIO_NAMESPACE;
-
-
-OCIO_ADD_TEST(RangeTransform, basic)
-{
-    OCIO::RangeTransformRcPtr range = OCIO::RangeTransform::Create();
-    OCIO_CHECK_EQUAL(range->getDirection(), OCIO::TRANSFORM_DIR_FORWARD);
-    OCIO_CHECK_EQUAL(range->getStyle(), OCIO::RANGE_CLAMP);
-    OCIO_CHECK_ASSERT(!range->hasMinInValue());
-    OCIO_CHECK_ASSERT(!range->hasMaxInValue());
-    OCIO_CHECK_ASSERT(!range->hasMinOutValue());
-    OCIO_CHECK_ASSERT(!range->hasMaxOutValue());
-
-    range->setDirection(OCIO::TRANSFORM_DIR_INVERSE);
-    OCIO_CHECK_EQUAL(range->getDirection(), OCIO::TRANSFORM_DIR_INVERSE);
-
-    range->setStyle(OCIO::RANGE_NO_CLAMP);
-    OCIO_CHECK_EQUAL(range->getStyle(), OCIO::RANGE_NO_CLAMP);
-
-    range->setMinInValue(-0.5);
-    OCIO_CHECK_EQUAL(range->getMinInValue(), -0.5);
-    OCIO_CHECK_ASSERT(range->hasMinInValue());
-
-    OCIO::RangeTransformRcPtr range2 = OCIO::RangeTransform::Create();
-    range2->setDirection(OCIO::TRANSFORM_DIR_INVERSE);
-    range2->setMinInValue(-0.5);
-    range2->setStyle(OCIO::RANGE_NO_CLAMP);
-    OCIO_CHECK_ASSERT(range2->equals(*range));
-
-    range2->setDirection(OCIO::TRANSFORM_DIR_FORWARD);
-    range2->setMinInValue(-1.5);
-    range2->setMaxInValue(-0.5);
-    range2->setMinOutValue(1.5);
-    range2->setMaxOutValue(4.5);
-
-    OCIO_CHECK_EQUAL(range2->getFileInputBitDepth(), OCIO::BIT_DEPTH_UNKNOWN);
-    OCIO_CHECK_EQUAL(range2->getFileOutputBitDepth(), OCIO::BIT_DEPTH_UNKNOWN);
-
-    range2->setFileInputBitDepth(OCIO::BIT_DEPTH_UINT8);
-    range2->setFileOutputBitDepth(OCIO::BIT_DEPTH_UINT10);
-
-    OCIO_CHECK_EQUAL(range2->getFileInputBitDepth(), OCIO::BIT_DEPTH_UINT8);
-    OCIO_CHECK_EQUAL(range2->getFileOutputBitDepth(), OCIO::BIT_DEPTH_UINT10);
-
-    OCIO_CHECK_EQUAL(range2->getMinInValue(), -1.5);
-    OCIO_CHECK_EQUAL(range2->getMaxInValue(), -0.5);
-    OCIO_CHECK_EQUAL(range2->getMinOutValue(), 1.5);
-    OCIO_CHECK_EQUAL(range2->getMaxOutValue(), 4.5);
-
-    range2->unsetMinInValue();
-
-    // (Note that the transform would not validate at this point.)
-
-    OCIO_CHECK_ASSERT(!range2->hasMinInValue());
-    OCIO_CHECK_EQUAL(range2->getMaxInValue(), -0.5);
-    OCIO_CHECK_EQUAL(range2->getMinOutValue(), 1.5);
-    OCIO_CHECK_EQUAL(range2->getMaxOutValue(), 4.5);
-
-    range2->setMinInValue(-1.5f);
-    OCIO_CHECK_EQUAL(range2->getMinInValue(), -1.5);
-    OCIO_CHECK_EQUAL(range2->getMaxInValue(), -0.5);
-    OCIO_CHECK_EQUAL(range2->getMinOutValue(), 1.5);
-    OCIO_CHECK_EQUAL(range2->getMaxOutValue(), 4.5);
-
-    OCIO_CHECK_ASSERT(range2->hasMinInValue());
-    OCIO_CHECK_ASSERT(range2->hasMaxInValue());
-    OCIO_CHECK_ASSERT(range2->hasMinOutValue());
-    OCIO_CHECK_ASSERT(range2->hasMaxOutValue());
-
-    range2->unsetMinInValue();
-    OCIO_CHECK_ASSERT(!range2->hasMinInValue());
-    OCIO_CHECK_ASSERT(range2->hasMaxInValue());
-    OCIO_CHECK_ASSERT(range2->hasMinOutValue());
-    OCIO_CHECK_ASSERT(range2->hasMaxOutValue());
-
-    range2->unsetMaxInValue();
-    OCIO_CHECK_ASSERT(!range2->hasMinInValue());
-    OCIO_CHECK_ASSERT(!range2->hasMaxInValue());
-    OCIO_CHECK_ASSERT(range2->hasMinOutValue());
-    OCIO_CHECK_ASSERT(range2->hasMaxOutValue());
-
-    range2->unsetMinOutValue();
-    OCIO_CHECK_ASSERT(!range2->hasMinInValue());
-    OCIO_CHECK_ASSERT(!range2->hasMaxInValue());
-    OCIO_CHECK_ASSERT(!range2->hasMinOutValue());
-    OCIO_CHECK_ASSERT(range2->hasMaxOutValue());
-
-    range2->unsetMaxOutValue();
-    OCIO_CHECK_ASSERT(!range2->hasMinInValue());
-    OCIO_CHECK_ASSERT(!range2->hasMaxInValue());
-    OCIO_CHECK_ASSERT(!range2->hasMinOutValue());
-    OCIO_CHECK_ASSERT(!range2->hasMaxOutValue());
-}
-
-#endif
